Moving list of games to GameListState instead of having data
