+++ src/ui/store/actions/game-list.tsx
@@ -1,6 +1,7 @@
 import { Action } from 'redux';
+import { GameInfo } from '../../../interfaces';
 
-export type GameListAction = MoveFocusedGame | SelectFocusedGame;
+export type GameListAction = MoveFocusedGame | SelectFocusedGame | UpdateGameList;
 
 export interface MoveFocusedGame extends Action {
   type: 'moveFocusedGame';
@@ -9,6 +10,11 @@ export interface MoveFocusedGame extends Action {
 
 export interface SelectFocusedGame extends Action {
   type: 'selectFocusedGame';
+}
+
+export interface UpdateGameList extends Action {
+  type: 'updateGames';
+  games: GameInfo[];
 }
 
 export function moveFocusedGame(delta: number): MoveFocusedGame {
@@ -21,5 +27,12 @@ export function moveFocusedGame(delta: number): MoveFocusedGame {
 export function selectFocusedGame(): SelectFocusedGame {
   return {
     type: 'selectFocusedGame',
+  };
+}
+
+export function updateGames(games: GameInfo[]): UpdateGameList {
+  return {
+    games,
+    type: 'updateGames',
   };
 }
